# Performance improvements for formatting in SqlString (caching)