--- a/lib/protocol/SqlString.js
+++ b/lib/protocol/SqlString.js
@@ -11,6 +11,7 @@ var charsMap   = {
   '\'': '\\\'',
   '\\': '\\\\'
 };
+var formatters = {};
 
 SqlString.escapeId = function escapeId(val, forbidQualified) {
   if (Array.isArray(val)) {
@@ -72,20 +73,20 @@ SqlString.arrayToList = function arrayToList(array, timeZone) {
 };
 
 SqlString.format = function format(sql, values, stringifyObjects, timeZone) {
-  values = values == null ? [] : [].concat(values);
+  if (values == null) {
+    return sql;
+  }
 
-  var index = 0;
-  return sql.replace(/\?\??/g, function(match) {
-    if (index === values.length) {
-      return match;
-    }
+  if (!(values instanceof Array)) {
+    values = [values];
+  }
 
-    var value = values[index++];
+  var formatter = formatters[sql];
+  if (!formatter) {
+    formatter = formatters[sql] = buildFormatter(sql);
+  }
 
-    return match === '??'
-      ? SqlString.escapeId(value)
-      : SqlString.escape(value, stringifyObjects, timeZone);
-  });
+  return formatter(SqlString, values, stringifyObjects, timeZone);
 };
 
 SqlString.dateToString = function dateToString(date, timeZone) {
@@ -148,6 +149,33 @@ SqlString.objectToValues = function objectToValues(object, timeZone) {
 
   return sql;
 };
+
+function buildFormatter(sql) {
+  var valuesIndex = 0;
+
+  function escapingReplacer(match) {
+    if (match.length === 1 && match !== '?') {
+      return charsMap[match];
+    }
+
+    var replacement = '(' + valuesIndex + ' < len ? ' + (
+      match === '?'
+        ? 's.escape(v[' + valuesIndex + '],so,tz)'
+        : 's.escapeId(v[' + valuesIndex + '])'
+    ) + ' : "' + match + '")';
+
+    ++valuesIndex;
+
+    return '" + ' + replacement + ' + "';
+  }
+
+  return new Function(
+    's,v,so,tz',
+    '"use strict";\n' +
+    'var len = v.length;\n' +
+    'return "' + sql.replace(/\?\??|["\r\n]/g, escapingReplacer) + '";'
+  );
+}
 
 function escapeString(val) {
   var chunkIndex = charsRegex.lastIndex = 0;