Improve performance by:
- Returning early if there are no values to use for formatting
- Not copying the `values` argument if it is already an Array
- Compiling and caching an optimized formatting function for each unique query string
#### Benchmark Results

| Query | Before (ops/sec) | This PR (ops/sec) | Speedup (`PR/before`) |
| --- | --- | --- | --- |
| no values | 1,950,229 | 52,944,361 | 27 |
| 1) simple query | 842,548 | 3,603,032 | 4.3 |
| 2) more values | 700,379 | 2,848,857 | 4.1 |
| 3) a lot of values | 419,132 | 950,663 | 2.3 |
| 4) long query string, few values | 748,445 | 3,538,848 | 4.7 |

See [this gist](https://gist.github.com/nwoltman/7de8311bb69b56bb0002265ad0e5df97) for the code.

The technique for compiling the formatting function was explained by [Felix](https://github.com/felixge) in his talk/post [Faster than C?](https://github.com/felixge/faster-than-c#examples). It's the same technique that [EJS](https://www.npmjs.com/package/ejs) is built on. Basically, this is how it works:

``` js
buildFormatter('SELECT * FROM `user` WHERE ?? = ?');
// returns a function that looks like this:
function formatter(SqlString, values, stringifyObjects, timeZone) {
  "use strict";
  var len = values.length;
  return "SELECT * FROM `user` WHERE " + (0 < len ? SqlString.escapeId(values[0]) : "??") + " = " + (1 < len ? SqlString.escape(values[1], stringifyObjects, timeZone) : "?") + "";
}
```

The `(index < values.length ? escape(values[index]) : 'placeholder')` is needed in case there are less values provided than placeholders.