@dougwilson 2016-05-29
> Compiling and caching an optimized formatting function for each unique query string

This is a significant addition and will definitely not land in 2.11.0, for through testing. One thing I notice right away is that it looks like it will increase indefinitely, which is a big problem that will need to be addressed before accepting for sure.

Would it be possible to split this PR apart so we can accept the other changes in the meantime, while the caching is discussed and iterated on?