@sidorares 2016-05-29
@nwoltman there was an attempt to factor out some non-core functionality so that others might benefit from optimisations like this one - there is a repo at https://github.com/mysqljs/sqlstring . It's currently very behind master, so as a step one we'll need to sync it. Just heads up, not saying that we should freeze non-factored-out sqlstring right now.