@dougwilson 2016-05-29
Funny you mention that module, as I was actually planning to sync it after I got @nwoltman s latest changed merged here :)