@dougwilson 2016-06-06
Alrighty, the code has been extracted to the 2.x line of the `sqlstring` module! I would suggest either moving this PR there or refactoring this PR to use that module in this code, whichever you feel is the most appropriate.