@nwoltman 2016-06-20
Since it's probably not a good idea to always use a cache that can grow indefinitely, I was thinking maybe [lru-cache](https://www.npmjs.com/package/lru-cache) should be used instead of a plain object and users can configure the cache size limit, let the cache be infinite, or not use the cache at all (I'm thinking this could be the default). Does that sound like the right way to go?