@sidorares 2016-06-20
@nwoltman sounds like the right way, I'm doing that in named-placeholders - https://github.com/mysqljs/named-placeholders/blob/89fd88865c39a6d825a0e2b54c5ce14e493c3876/index.js#L75