@dougwilson 2016-06-20
@nwoltman, remember that `lru-cache` will not install on Node.js 0.6 or 0.8, so at least for a while, we cannot use that as a dependency in this module.