@sidorares 2016-06-20
@dougwilson is that new (lru-cache not compatible with 0.8)? I definitely used it with 0.8 previously, but I dropped 0.8 support some time ago and don't now for sure now