@dougwilson 2016-06-20
@sidorares, the dependency definitions are using the `^` symbol, which does not work out-of-the-box on the versions on npm is 0.6, 0.8, and early 0.10 versions.