@nwoltman 2016-06-20
Okay, how about having an internal cache class with no limits that just uses a plain object and allowing users to pass in their own cache instance if they want to limit the cache size ([like EJS](https://github.com/mde/ejs#caching))?