@dougwilson 2016-06-20
@nwoltman, that could work, as long as it is off by default. And at that point, why not have it so they have to pass in their own cache instance if they ever want it to be on? We should very clearly define what this cache class's contract is, including all methods, how we call those methods (as methods or functions), what happens on throws, how callbacks work etc.