@nwoltman 2016-06-20
The benefits of having a simple, internal infinite cache are:

1) It's better for performance and memory.

Formatting `'SELECT * FROM user WHERE ?? = ?'` yields:
```
current x 2,158,570 ops/sec ±0.59% (91 runs sampled)
with simpleObject x 4,031,955 ops/sec ±0.52% (95 runs sampled)
with lru-cache x 2,614,982 ops/sec ±0.67% (95 runs sampled)
```
(although formatting `'UPDATE ?? SET ?? = ?, ?? = ?, ?? = ? WHERE id = ?'` yields:
```
current x 566,961 ops/sec ±1.66% (84 runs sampled)
with simpleObject x 1,066,110 ops/sec ±0.54% (93 runs sampled)
with lru-cache x 919,604 ops/sec ±0.70% (95 runs sampled)
```
)

2) Users who don't have memory concerns won't have to deal with another dependency.

3) The internal cache can be shared among connections/pools, which will reduce memory and garbage collector invocations.