@dougwilson 2016-06-20
Yes, but people barely read the documentation as it is. You're just never going to convince me that we will, but default, have an unbounded cache in a core module that is used in servers that are never restarted.