@nwoltman 2016-06-20
Sorry for not specifying, I agree it definitely shouldn't be on by default. I was thinking it would be off by default and then:

```js
// Pool/connection config
{formatCache: cacheInstance}, // recommended option
{formatCache: true} // infinite cache
```