@dougwilson 2017-01-14
I was just trying to comb through the issues and see we were still talking in this while the `SqlString` was split apart to share with `mysql2` into https://github.com/mysqljs/sqlstring

So besides the conversation above, on the technical side should the caching be in that module, or update the PR to move around how it's integrated into this module?