@nwoltman 2017-01-28
It would probably be nice to have the caching in `SqlString` so that `mysql2` could also benefit from it, but I think it makes the most sense to have the caching in this module because `SqlString` is a simple library of escaping functions and shouldn't be concerned with configuration or keeping any sort of state.