@nwoltman 2018-01-22
I'm going to close this because the performance/memory trade-off probably isn't worth it and escaping query values is already pretty fast so it's unlikely to be a bottleneck for anyone.